LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_cached-expected.txt

Description

This test validates that a 304 Not Modified resource appears in the buffer.


PASS window.performance is defined 
PASS window.performance.getEntriesByType is defined 
FAIL There should be two entries assert_equals: There should be two entries expected 2 but got 0
